+++ Constraid/ExpandFromSize.swift
@@ -5,25 +5,39 @@
 #endif
 
 extension ConstraidView {
-    open func expand(fromWidthOf item: Any?, constant: CGFloat = 0.0,
-                     multiplier: CGFloat = 1.0, priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) {
+    @discardableResult
+    open func expand(fromWidthOf item: Any?,
+        constant: CGFloat = 0.0,
+        multiplier: CGFloat = 1.0,
+        priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired
+        ) -> ConstraidConstraintCollection {
 
         self.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
+        let collection = ConstraidConstraintCollection([
             NSLayoutConstraint(item: self, attribute: .width,
-                               relatedBy: .greaterThanOrEqual, toItem: item, attribute: .width,
-                               multiplier: multiplier, constant: constant, priority: priority)
+                relatedBy: .greaterThanOrEqual, toItem: item,
+                attribute: .width, multiplier: multiplier,
+                constant: constant, priority: priority)
             ])
+        collection.activate()
+        return collection
     }
 
-    open func expand(fromHeightOf item: Any?, constant: CGFloat = 0.0,
-                     multiplier: CGFloat = 1.0, priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) {
+    @discardableResult
+    open func expand(fromHeightOf item: Any?,
+        constant: CGFloat = 0.0,
+        multiplier: CGFloat = 1.0,
+        priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired
+        ) -> ConstraidConstraintCollection {
 
         self.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
+        let collection = ConstraidConstraintCollection([
             NSLayoutConstraint(item: self, attribute: .height,
-                               relatedBy: .greaterThanOrEqual, toItem: item, attribute: .height,
-                               multiplier: multiplier, constant: constant, priority: priority)
+                relatedBy: .greaterThanOrEqual, toItem: item,
+                attribute: .height, multiplier: multiplier,
+                constant: constant, priority: priority)
             ])
+        collection.activate()
+        return collection
     }
 }