+++ ConstraidTests/ExpandFromSizeTests.swift
@@ -2,15 +2,15 @@ import XCTest
 import Constraid
 
 class ExpandFromSizeTests: XCTestCase {
-
     func testExpandFromWidthOf() {
         let viewOne = UIView()
         let viewTwo = UIView()
 
         viewOne.addSubview(viewTwo)
-        viewOne.expand(fromWidthOf: viewTwo, constant: 10.0, multiplier: 2.0, priority: 500)
+        let constraints = viewOne.expand(fromWidthOf: viewTwo, constant: 10.0, multiplier: 2.0, priority: 500)
         let constraint = viewOne.constraints.first! as NSLayoutConstraint
 
+        XCTAssertEqual(constraints, viewOne.constraints)
         XCTAssertEqual(constraint.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraint.firstAttribute, NSLayoutAttribute.width)
         XCTAssertEqual(constraint.relation, NSLayoutRelation.greaterThanOrEqual)
@@ -26,9 +26,10 @@ class ExpandFromSizeTests: XCTestCase {
         let viewTwo = UIView()
 
         viewOne.addSubview(viewTwo)
-        viewOne.expand(fromHeightOf: viewTwo, constant: 10.0, multiplier: 2.0, priority: 500)
+        let constraints = viewOne.expand(fromHeightOf: viewTwo, constant: 10.0, multiplier: 2.0, priority: 500)
         let constraint = viewOne.constraints.first! as NSLayoutConstraint
 
+        XCTAssertEqual(constraints, viewOne.constraints)
         XCTAssertEqual(constraint.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraint.firstAttribute, NSLayoutAttribute.height)
         XCTAssertEqual(constraint.relation, NSLayoutRelation.greaterThanOrEqual)